Improve scaling of canvas relative to DPI
To achieve crisp graphics we need to deal with the gap between logical
and physical resolution of modern high-resolution screens.

The basic idea is to set the canvas dimensions to the physical
resolution of the device, but scale it down to the logical size of the
viewport.

A common reference is
https://www.html5rocks.com/en/tutorials/canvas/hidpi/ though it seems to
have been edited over the years and may continue to change.