Make distance/speed less volatile for bad frame rates
The game originally tried to run a game loop at 50 FPS and does not look
at actual time between frames when integrating speed to
position/distance travelled.

With this change, the actual delta between frames is taken into account
to determine how many "standard intervals" should be integrated over.
For example, if one frame was dropped (for an interval of 40 ms), double
the amount is added to the distance travelled.