Fix wheelchair setting
The sensitivity adjustment for the wheelchair flag had not been
reapplied at the correct place after the switch to directions expressed
in radians.

This applies the senstivity adjustment to the horizontal deviation
before computing the angle. Sensitivity is set to one of two values in
the config based on the wheelchair flag, it would be possible to let
users choose from a continuous range instead.